Adjust refactor: map.go back into maptest; amalgamate map_test.go and main_test.go; have map.go export a test table
